requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4

# Scraping
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.28.1
selenium==4.27.1
webdriver-manager==4.0.2

# Data Processing
pandas==2.2.3
numpy==2.2.1

# AI Assistance (OpenRouter-compatible)
openai==1.58.1

# Utilities
python-multipart==0.0.20
aiofiles==24.1.0
